# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pragma-api",
    "description": "0D Master API for vaults",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://{environment}.0d.finance/api/v1",
      "variables": {
        "environment": {
          "default": "devnet",
          "enum": [
            "devnet",
            "production"
          ]
        }
      }
    }
  ],
  "paths": {
...
    "/vaults/{vault_id}/caps": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "operationId": "get_vault_caps",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current values vs configured limits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultCapsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Vault not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/vaults/{vault_id}/liquidity": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "operationId": "get_vault_liquidity",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vault liquidity summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultLiquidityResponse"
                }
              }
            }
          },
          "404": {
            "description": "Vault not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/vaults/{vault_id}/liquidity/curve": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "operationId": "get_vault_slippage_curve",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vault slippage curve",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultSlippageCurveResponse"
                }
              }
            }
          },
          "404": {
            "description": "Vault not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/vaults/{vault_id}/liquidity/simulate": {
      "post": {
        "tags": [
          "Vaults"
        ],
        "operationId": "simulate_vault_liquidity",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiquiditySimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Liquidity simulation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquiditySimulateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "404": {
            "description": "Vault not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/vaults/{vault_id}/nav/latest": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "operationId": "get_vault_nav_latest",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest NAV report & deltas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultNavLatestResponse"
                }
              }
            }
          },
          "404": {
            "description": "Vault not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/vaults/{vault_id}/timeseries": {
      "get": {
        "tags": [
          "Vaults"
        ],
        "operationId": "get_vault_timeseries",
        "parameters": [
          {
            "name": "vault_id",
            "in": "path",
            "description": "Vault identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "description": "Vault-level metric to return",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "tvl"
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "Time period",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "7d"
          },
          {
            "name": "currency",
            "in": "query",
            "description": "Currency for values",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "USD"
          }
        ],
        "responses": {
          "200": {
            "description": "Vault timeseries data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultTimeseriesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Vault not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CapItem": {
        "type": "object",
        "required": [
          "name",
          "current",
          "limit",
          "unit"
        ],
        "properties": {
          "current": {
            "type": "number",
            "format": "double"
          },
          "limit": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          }
        }
      },
...
      "InstantLiquidity": {
        "type": "object",
        "required": [
          "supported",
          "est_slippage_bps",
          "cap_remaining"
        ],
        "properties": {
          "cap_remaining": {
            "type": "string"
          },
          "est_slippage_bps": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "supported": {
            "type": "boolean"
          }
        }
      },
      "LiquiditySimulateRequest": {
        "type": "object",
        "required": [
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "string"
          }
        }
      },
      "LiquiditySimulateResponse": {
        "type": "object",
        "required": [
          "amount",
          "scheduled"
        ],
        "properties": {
          "amount": {
            "type": "string"
          },
          "instant": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InstantLiquidity"
              }
            ]
          },
          "scheduled": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduledWindow"
            }
          }
        }
      },
...
      "ScheduledWindow": {
        "type": "object",
        "required": [
          "window",
          "expected_nav_date"
        ],
        "properties": {
          "expected_nav_date": {
            "type": "string"
          },
          "max_without_delay": {
            "type": [
              "string",
              "null"
            ]
          },
          "window": {
            "type": "string"
          }
        }
      },
      "SlippagePoint": {
        "type": "object",
        "required": [
          "amount_usd",
          "slippage_bps"
        ],
        "properties": {
          "amount_usd": {
            "type": "string"
          },
          "slippage_bps": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "TimeseriesPoint": {
        "type": "object",
        "required": [
          "t",
          "v"
        ],
        "properties": {
          "t": {
            "type": "string"
          },
          "v": {
            "type": "string"
          }
        }
      },
...
      "VaultCapsResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CapItem"
            }
          }
        }
      },
...
      "VaultLiquidityResponse": {
        "type": "object",
        "required": [
          "is_liquid",
          "withdraw_capacity_usd_24h",
          "deposit_capacity_usd_24h"
        ],
        "properties": {
          "as_of": {
            "type": [
              "string",
              "null"
            ]
          },
          "deposit_capacity_usd_24h": {
            "type": "string"
          },
          "is_liquid": {
            "type": "boolean"
          },
          "policy_markdown": {
            "type": [
              "string",
              "null"
            ]
          },
          "withdraw_capacity_usd_24h": {
            "type": "string"
          }
        }
      },
...
      "VaultNavLatestResponse": {
        "type": "object",
        "required": [
          "date",
          "aum",
          "var_since_prev_pct",
          "apr_since_prev_pct",
          "report_url"
        ],
        "properties": {
          "apr_since_prev_pct": {
            "type": "number",
            "format": "double"
          },
          "aum": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "report_url": {
            "type": "string"
          },
          "var_since_prev_pct": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "VaultSlippageCurveResponse": {
        "type": "object",
        "required": [
          "is_liquid",
          "points"
        ],
        "properties": {
          "is_liquid": {
            "type": "boolean"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SlippagePoint"
            }
          }
        }
      },
...
      "VaultTimeseriesResponse": {
        "type": "object",
        "required": [
          "metric",
          "timeframe",
          "points"
        ],
        "properties": {
          "metric": {
            "type": "string"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeseriesPoint"
            }
          },
          "timeframe": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}